Encapsulated repeated code into a trait
